requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
discord.ext.context==0.1.7
discord.py==1.7.3
idna==2.10
multidict==5.1.0
python-dotenv==0.18.0
requests==2.25.1
selenium==3.141.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
